Global Library support for multiple instances.
Enforce a depot syntax path ending of "/..."
Remove 'beta' from Global Library configuration.
Generate a unique workspace to sync multiple Libraries. JENKINS-53922
