+++ src/main/resources/org/jenkinsci/plugins/p4/scm/GlobalLibraryScmSource/help-path.html
@@ -1,4 +1,4 @@
 <div>
 	<p>A Perforce Depot path to the library.</p>
-	<p>For example: <code>//depot/myLibrary</code></p>
+	<p>For example: <code>//depot/myLibrary/...</code></p>
 </div>
